feat: implement GetServerSideProps
