+++ next.config.js
@@ -9,3 +9,4 @@ const nextConfig = {
 };
 
 module.exports = nextConfig;
+
