+++ src/action.ts
@@ -1,6 +1,6 @@
 import { createAction } from '@reduxjs/toolkit';
 
-import { Starship } from './pages/App';
+import { Starship } from './pages';
 
 export const updateSearchTerm = createAction<string>('updateSearchTerm');
 export const updateItemsPerPage = createAction<number>('updateItemsPerPage');
